ui: rm selfsigned.crt mnt from graphql cont
Assumption: the docker-compose.yaml file is
for local development, only.

yarn services:start failed for me with

  ERROR: for graphql  Cannot create container for service graphql: invalid
  mount config for type "bind": bind source path does not exist:
  /home/jp/dev/opstrace/packages/app/secrets/cert.crt

I thought about this. It tries to mount to

  /etc/ssl/certs

which is a place to add trust anchors (root CA
certificates -- or the public part of a self-signed
cert) to a trust store; to be picked up by
popular software. That only makes sense when
somewhere across the software stack there is
a service that exposes a corresponding cert
(together with the private key). Which I
couldn't find traces of via grep. So, I expected
things to simply work when removing the mount.

yarn services:start -> works
yarn server:start -> works
yarn client:start -> works

Let's see if this breaks things further down
the line. Then we have to add proper instructions.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>