chore: await axios.request() (err handling fix)
Suspicious:

    return axios.request(deleteRequest);

Especially being called towards the _end of
the program_ this might result in the
HTTP response not being waited for at
all.

If it is waited for/received then, depending on
other weaknesses in the program, it's unclear
which part of the program is supposed to
catch and handle an unexpected HTTP response.

Root problem: forgetting to put the `await`
keyword where it has to be put.

Carefully putting the `await` keyword where
it has to be put is on the developer --
tooling rarely helps. See

https://github.com/eslint/eslint/issues/13567

When missing an `await` keyword then
the value propagating through code is
often not what we think it is. With a
weak return type signature such as
`Promise<any>` the _actual_ return value
might in fact be a Promise to a Promise.
This is dangerous territory; I was peeking
into

https://github.com/microsoft/TypeScript/issues/27711

two things are clear from this issue:

- TypeScript is in flux w.r.t. situations
  involving things like Promise<Promise<T>>

- I don't even want to know about this topic

The Delete() method is called here:
https://github.com/opstrace/opstrace/blob/f610692c86d0da1d16165c89c22d814594940523/lib/dns/src/index.ts#L208
via redux-saga. Redux-saga has its own
fancy interaction with async functions,
and its own error handling paradigms.

I assume that as of the missing `await` and
the weak return type signature the Delete()
method might in fact have been subject to a
race condition, leaving a "dangling thing"
behind with no proper error handling mechanism
looking at that thing anymore after submission.

That is, it's likely that sometimes the program
either didn't wait for receiving the HTTP response
or didn't have an error handler taking action
when a bad response was received.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>