makefile: test-remote: add artifact path concept
So far, test-remote did not need to leave
behind file artifacts. The current working
directory in containerized execution was
in the container, not exposed.

This will stay as-is, but now there
is another directory mounted into
the container at /test-remote-artifacts.
At the same time, the test runner code
has a new concept of TEST_REMOTE_ARTIFACT_DIRECTORY
which defaults to `.` -- that is, in uncontainerized
execution the test runner writes artifacts to
the current working directory .

In containerized execution the test runner
writes artifacts to /test-remote-artifacts
which is where the OPSTRACE_BUILD_DIR is
mounted.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>