chore: lib/dns: handle expired authentication proof
This resolves issue #53.

When the DNS service responds to and HTTP
request with a 401 HTTP response then
this means that the authentication proof
(set of credentials) presented in the request
was invalid.

A common reason for that is authentication
proof / credential expiry. Therefore, a valuable
course of action is to obtain a fresh set of
credentials, by going through another login
process.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>