config: Rename 'graphql' API to 'config' API, accept tenant via bearer (#351)
rename service: The prior `graphql` name was quickly confusing for everyone involved because it doesn't actually serve graphql. Instead it's just an implementation detail that it acts as a frontend to the graphql service. Since the service handles creating, updating, and deleting credentials/exporters entries in graphql, and might be extended to support changing other settings in the future, it felt like `config` would be a reasonable name for the service, at least for now.

change tenantname handling: The config service previously was "locked" on startup to a single tenant via CLI argument, with instances running in each tenant namespace. However, testing quickly showed that the tenant namespaces don't have access to the hasura/graphql secret. To resolve this, two changes are made to the config service:
- The `--tenantname` argument is removed and the service is no longer "locked" to a single tenant on startup
- The service instead reads the tenant name from the (validated) bearer token in HTTP requests
- The `middleware` package has an API call added for retrieving the validated tenant name from a bearer token
With these changes, the service can instead be run from the `application` namespace, where it will have access to Hasura secrets. It would have been a bit redundant to run the service in each tenant namespace anyway. The alternative option would have been to copy the hasura secret into each per-tenant namespace, which felt much less desirable.

Signed-off-by: Nick Parker <nick@opstrace.com>