chore: add local retry for SQLInstance creation
If we move too quickly with creating the instance after a peering request,
we'll get an error. Usually the second retry will succeed.
Would be nice to find a clean way to wait for the peering to be complete
but I couldn't find anything to help me with that, so we'll just keep retrying
for 2 mins.

Signed-off-by: Mat Appelman <mat@opstrace.com>