controller: Restart cloudwatch exporter automatically when config content changes
Adds a bash sidecar to the cloudwatch pod that checks the config file in a loop. When a change is detected, the sidecar exits, causing the exporter to restart with the new config. A bit hacky but should get the job done, and better than leaving the pod to ignore config changes.

Inspired by https://github.com/jimmidyson/configmap-reload , which unfortunately doesn't work for us since the cloudwatch exporter doesn't have a reload nor restart HTTP endpoint.

Signed-off-by: Nick Parker <nick@opstrace.com>