installer: set tenant_api_authenticator_pubkey_set_json
This sets the new key

   tenant_api_authenticator_pubkey_set_json

on the controller config, for communicating
the serialized set of public keys for the
authenticator instances.

This also adjusts the cluster config types,
and adjusts the name of the environment
variable which the autheticator instances
expect to be configured with.

This does not solve the upgrade problem
and drops support for the controller
config parameter `data_api_authn_pubkey_pem`.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>