bump: prometheus, prometheus-operator, prometheus CRDs (fixes #478)
The root problem was that the tenant prometheus should talk directly to cortex because it doesn't have an auth token to communicate via `cortex-api`. The `system` instance meanwhile worked fine because it has a token available.
However talking to cortex directly required adding the `X-Scope-OrgID` tenant header to requests.
And then THIS required updating the prometheus stack (prometheus-operator, CRDs, and prometheus itself) because the `remoteWrite.headers` option was only added very recently and each component needed an update to support it.

After all of these updates, I was able to get working data in Grafana etc again, however I did need to manually delete the `prometheuses` CRD so that it would be recreated automatically by the new controller.

Some other cleanup along the way:
- Some hardcoded docker images have been moved out to `docker-images.json`
- The process to update prometheus CRDs has been updated to allow generating against arbitrary versions of the prometheus operator (as of writing there were no CRDs generated for operator 0.46.0 yet), and to avoid editing the files by hand.

Signed-off-by: Nick Parker <nick@opstrace.com>