controller: Generate controller graphql API alongside other copies
Instead of copying the file from packages/app in the controller build, just update both copies at once.
This avoids potential issues around the controller graphql code falling out of date vs the app graphql code.

As a future improvement, the graphql schema itself could be moved into a separate package. However it depends on Hasura, so we'd probably need to bring that along as well, at which point it starts entering "can of worms" territory in terms of updating build/dev tooling.

Signed-off-by: Nick Parker <nick@opstrace.com>