installer: mark system-tenant-api-auth-token as managed
The `system-tenant-api-auth-token` secret lacked an `opstrace` annotation, so the controller would then log a notice about it not being a known resource. Passing it via a `ResourceCollection` automatically configures the annotation. We also mark the secret as immutable since it's not explicitly included among the controller's expected resources.

As a minor change, the notice message is updated to include the type of the resource being logged (`Secret` in this example).

Signed-off-by: Nick Parker <nick@opstrace.com>