controller: Support Azure credentials and metrics exporters (#556)
Adds an `azure-service-principal` credential type and `azure` exporter type. This was tested manually using a freebie Azure account and seems to work fine. Much of the time here was spent just figuring out how to get the credential configured in the Azure Portal UI (hint: Your Application will NOT appear in the Role Assignments dialog until you search for its name in the search bar)

For the exporter itself, this is using a patched build that has support for envvar credentials - whereas the stock build requires embedding the credentials in the yaml configuration alongside the list of metrics to fetch. The addition of envvar support is pretty straightforward and allows us to avoid dealing with mangling the yaml configuration file. See also: https://github.com/RobustPerception/azure_metrics_exporter/pull/101

Signed-off-by: Nick Parker <nick@opstrace.com>