ui: Tweaks to K8s prometheus integration, add delete command
At this point, the prometheus integration worked with getting metrics out of an on-prem K3s cluster that I'm running in the office here and into a test opstrace instance.
The configuration could be improved - for example in my case I wasn't able to get node collection working from the K3s nodes and only a few pods were emitting metrics at the hardcoded `/metrics` endpoint, but for an MVP this should be okay for now.

Signed-off-by: Nick Parker <nick@opstrace.com>