ci: do not install playwright for root workspace
Best fix for instabilities around #696
is to not try to install playwright where
it is not needed. It is in fact only needed
for the test-remote project.

When installing dependencies with yarn before
linting, it's absolutely not needed. It's also
not needed for the CLI build, for the controller
build etc.

However, it is difficult or impossible to find
a clean, explicit way to skip installing
a certain package with yarn -- it goes deep into
the file system, discovers all package.json
files and then installs _all_ dependencies.

The --ignore-optional is one way to maybe
find a hacky way to skip installing playwright
before linting the codebase.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>