cli: add tenant name validation
See issue #710

Add this on top of 'schema2', pretending
that this was here implicitly before:

tenant names not matching this regex
did not lead to a working system.

this change only makes that behavior
explicit, and I think it is very fine
to not introduce a new config schema
version for it.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>