chore: remove .vscode dir
In a previous commit, we deliberately decided
for using a workspace config file, and against
a local .vscode directory.

Previous commit message:

I could not find a way to configure multiple
folders per workspace when using the .vscode
directory. That makes sense: the .vscode directory
defines the current directory (that it is in)
as a workspace.

Instead, we need to have a multi-root
workspace which seemingly requires the
workspace settings file.

Ref: https://code.visualstudio.com/docs/editor/multi-root-workspaces

microsoft/vscode-go#2935
golang/go#32394

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>