ui: Update promtail configuration following testing (#688)
Adds a config option for the user to specify whether they are running dockerd or cri/containerd. This is needed because the two have totally different log formats, otherwise we could have gotten away with supporting both in one universal config. Its conceivable that we could add some kind of autodetect logic to the promtail pods pre-startup, but it's probably simpler/more reliable to just ask.

Status as of this commit:
- EKS dockerd cluster: working great, logs from all pods are showing up in loki
- onprem containerd cluster: partially working, logs from only the promtail pods themselves are showing up in loki (makes me think there's an issue with pod discovery, whereas log scraping itself is working fine)

Signed-off-by: Nick Parker <nick@opstrace.com>