ui: Updates to promtail configuration following testing
Adds a config option for the user to specify whether they are running dockerd or cri/containerd. This is needed because the two have totally different log formats, otherwise we could have gotten away with supporting both in one universal config. Its conceivable that we could add some kind of autodetect logic to the promtail pods pre-startup, but it's probably simpler/more reliable to just ask.

Updates labeling configuration across prometheus/promtail so that they match each other, and to remove dependency on pods being labeled a certain way.

Signed-off-by: Nick Parker <nick@opstrace.com>