controller: Switch to polling method for credentials/exporters (#615) (#758)
* controller: Switch to polling method for credentials/exporters (#615)

We're seeing a CI failure in test-remote where all of the exporter tests will sometimes fail. When this happens, it looks like the controller is effectively failing to see that anything is supposed to be deployed. Meanwhile all non-exporter tests would pass fine. The failures so far have mainly been on GCP, which gives the impression that it's a GCP-specific networking/environment issue, but hopefully one that can be worked around on behalf of users. That said, there have also been one or maybe two failures with the same symptoms on AWS.

The credential/exporter sync is the only place in the controller that's using GraphQL subscriptions, so I'm thinking they are related to the problem. Perhaps the controller is rarely failing to connect at all (again maybe due to a GCP thing), which then results not seeing/deploying the credentials/exporters? The issue itself is hard to repro, so this is a bit of an experiment to see if the issue goes away on its own (or alternatively, gets worse?).

PS: It's feeling likely that credentials/exporters will be getting combined and folded into the single Integrations table/query at some point, which would result in a single query loop rather than two separate loops. This would make using a basic poll an (even) easier sell.

Signed-off-by: Nick Parker <nick@opstrace.com>

* controller: Update generated schemas following subscription call removal

Signed-off-by: Nick Parker <nick@opstrace.com>