controller: Configure ingesters to not block on readiness (of other ingesters)
The ingester readiness check is based on the ring status of all ingesters.
As a result you can end up with e.g. ingester-1 refusing to deploy because ingester-5 is unhappy.
The StatefulSet will then refuse to deploy ingester-5 because it wants ingester-1 to be ready first.

This breaks the dependency cycle by allowing ingesters to deploy (and/or scale out) in parallel. Note that it does not affect the rollout of config changes, which will remain sequential.

Signed-off-by: Nick Parker <nick@opstrace.com>