looker: rename logstream-gen to looker
logstream-gen was the working title
when I started building a tool with
limited scope to synthetically generate
a log stream to push into Loki. It
seems like we've settled on the
name looker for the wider scope thing
(including readback and metrics mode)
and this should be reflected in code.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>