controller: Configure inflight push request limits (#986)
By default, the ingester and distributor have an unlimited queue for pending push requests. This can lead to OOMing pods under heavy load (ingesters: lots of new series, distributors: large payloads with lots of labels).

This is a particular issue with ingesters, which lose in-memory state and take a while to recover, during which time cortex as a whole can become degraded.

These limits should reduce the likelihood of OOMing regardless of load. The comments give some justification around the chosen values, for more background see #966.

Signed-off-by: Nick Parker <nick@opstrace.com>