cli: destroy: do dns service teardown first
The AWS route53PurgeZonesForDnsName() for
example must come _after_ that, otherwise
the doesOpstraceIoDNSNameExist() will always
result in `false` (name resolution fails, of
course, because the authoritative name servers
are gone).

I've used this opportunity to move this
DNS service teardown part into the AWS-GCP-common
index.ts, and that also gets rid of the
problem where the DNS service login is
triggered minutes after invoking the
`destroy` operation (with a browser
window popping up late in the game). See

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>
#353