cli: destroy: doesOpstraceIoDNSNameExist() w/ DNS client
Rework doesOpstraceIoDNSNameExist() to use an
actual DNS client, specifically the DNS client
in the NodeJS stdlib.

It's too difficult to do get proper DNS resolution
error detection done with the help of an HTTP
client :-).

(TCP connect timeout errors would need to be expected
as a definite signal for name resolution having
worked, but the TCP connect having failed, and
reliable detection of that based on error message
string matching is getting too dirty).

Even with the DNS client appraoch, there is an
ambiguous situation here where we use the
ESERVFAIL error to detect a situation where
an NS record set _actually_ exists.

It is absolutely conceivable that this detection
method yields false positives; namely when
the DNS servers for the *.opstrace.io zone
misbehave (which I believe should be absolutely
unlikely, but I might miss something here).

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>