ci(upgrades): dissociate from test-core.sh
The deploy-testremote-teardown code path relies on docker-in-docker
magic. That was never required for the upgrade tests until we tried
to consolidate how we run the tests in both pipelines.

The previous attempt didn't work but we'll try again in the future.

For now, this commit reverts the code sharing between the test
pipelines and adds a script to set up the requirements for one of
the tests.

Signed-off-by: Simão Reis <sreis@opstrace.com>