ci(upgrades): check ring status in all loki distributor pods
It takes a while for the membership ring status to propagate to
all distributors so we need to wait until all the pods report
a correct ring status.

Also, GCP does not seem to like 'kubectl get raw' so we need to
set up port forwarding to the distributor pods to be able to
curl the endpoints.

Signed-off-by: Simão Reis <sreis@opstrace.com>