test-remote: test_ui_api: retry browser-based login
There are a bunch of timing fluctuations and
transient issues that are expected and that we
cannot really do anything about. There are also
instabilities that we could do something about,
I suppose. This here is a hammer that is supposed
to reduce the impact of all of these instabilities
at once. Note that this has the cost of us not
feeling the pain of the instabilities in the UI
login flow that we impose on our users as of
non-robust code in both, client (browser) and
server (express app).

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>