makefile: rename to OPSTRACE_KUBECFG_FILEPATH_ONHOST
This environment variable is meant to hold a path
to a kubectl configuration _file_ (not directory)
on the _host_ (i.e., this can be used as the left-
hand side of `docker run -v left:right`.

Adjust the -e KUBECONFIG environment variable
setters to point to _that file_ which is mounted
into the container at /kubeconfig (now /kubeconfig
should really always be a file, and never a
directory).

Docs for the KUBECONFIG env var:

"If the KUBECONFIG environment variable does exist,
kubectl uses an effective configuration that is the
result of merging the files listed in the
KUBECONFIG environment variable."

from https://kubernetes.io/docs/concepts/configuration/
organize-cluster-access-kubeconfig/#the-kubeconfig-environment-variable

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>