ci: bk pipeline.yal: introduce OPSTRACE_ARTIFACT_DIR env var
Make it easier for shell scripts to copy data there.
When you want to use that, note:
this carries an absolute path on the host file system
on the /tmp directory. That is, this works in all
containers that have a -v /tmp:/tmp which is what
we do for practically all containers launched in CI.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>