ci: Use age, not count, when deciding whether to delete CI buckets
The current logic avoids deleting the 20 most recent buckets.
However if we have a lot of concurrent CI jobs (>5), this can delete buckets that are being used by live tests.

To avoid this, we can explicitly check that the bucket was created at least 6 hours ago before we start trying to delete it.
This assumes that CI runs generally don't take longer than 6 hours.
This is done via a short stdlib-only python script because doing it in shell script gets pretty messy.

Signed-off-by: Nick Parker <nick@opstrace.com>