ui: consolidate interaction with /_/auth/status
This commit has a lot to it. See code comments.

At the high level, this is about useAxios
from axios-hooks not supporting retries.

So, this replaces axios-hooks with more raw
axios usage; with custom retrying logic on top
of it.

This then corresponding also reworks react
state management and while doing so removes
a set of state management problems.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>