cli: gcp subnet destroy: log error detail
The previous error handler logic allowed
for 404 API errors during DELETE to be swallowed,
hiding the reason for why an infinite loop was
entered:

There may be an asymmetry between
lookup and delete as of a region mismatch
when the region of the looked up network was
not matching the region used in the DELETE
request.

Always log error detail unless being
completely sure... so, that debug log
statement is important.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>