looker: a double-ended queue to rule them all
This changes paradigms for

- throttling
- write concurrency
- fragment generation task management

This introduces the "task queue" concept
that I like very much when thinking about
concurrency challenges.

At the heart of the write phase of a cycle
there may be millions of time series
representations, and for each one we may
want to write many fragments. And to
complicate things we sometimes may want
to bundle fragments from various time
series into the same binary push
message. All this needs to be organized
properly, and so far it kind of was.

What changed recently is that not all of
the work items are 'ready yet' when
the code looks at them (not ready for
log/metric sample generation). Why?
Because of the throttling mechanism
which comes with wall time coupling:
synthetic time must not overtake wall
time.

So far, when the code picked a time
series object to generate data for
but found that it's not yet ready
it fell into a lazy mode: it kept
looking at the work item periodically,
staying _idle_ until it became ready.

That could have taken seconds or minutes.

So. A worker / execution unit / actor
was sitting idle, waiting for just a
single time series object to become
'ready'.

While potentially tens of thousands or
much more _other_ work items were
ready to be processed.

This was not only "inefficient" per se.
For certain configurations where the
number of time series objects was much
larger than the number of concurrent
writers this could stall the machinery,
effectively bringing it to a halt.

This demanded a change to the concurrency
architecture; a method that allows actors
to quickly iterate through all remaining
jobs to find work that can be done _now_.

Another complication: so far, indivdual
actors (functions / executors) were supposed
and guaranteed to only act on a subset of
all time series objects. With the idea
goal of processing work as quickly as
possible of course all actors need to
be able to see all available work, and
coordinate among themselves.

So, use something like a ring buffer
(think: work queue) where each item
corresponds to a potential item of work,
a candidate.

Each executor can pop off an item
at the right end, and inspect it.

If it's not yet ready to be processed,
it can be added back to the ring buffer
on the left-hand side.

With potentially millions of objects
on this work queue / ring buffer it is
important to use an approriate data
structure. A JavaScript array has
slow time complexity for preprending
items.

A natural choice is a double-ended queue
for the following time complexity
characteristics:

- O(1) removal at right end
- O(1) insertion at left end

Now the actors can iterate spin
the items over the ring buffer
at 'full speed' (CPU-bound),
which is a good choice for
the scenario where there are orders
of magnitude more items than actors.

When items are 'done with', they
are removed from the ring buffer.

When the count of candidate items
on the ring buffer drops below the
number of actors then of course
it does not make sense to busy-
spin them around -- then we can fall
back to the polling mode where
an individual item is watched by a
single actor in a periodic fashion.

With this change I noticed that the
'stop criterion' (for when each
time series is really 'done' with
writing in the current cycle) should
be implemented deeper in the core
where each time series object is
actually processed.

That resulted in the following change:
if one or more fragments are still to
be written for the currently held
time series object (after one was just
written!) then simply put it back onto
the work queue / ring buffer. That
makes things more easy to reason
about.

Btw, I could only confidently make this
change because

https://github.com/invertase/denque

looks real good (mature, well-documented,
fast).

I think this commit still has bugs but
the conceptual change is largely there.

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>