controller: manage tenant certs separately (#1371)
* controller: manage tenant certs separately

The controller currently creates one certificate/secret up-front, and then copies the secret to the tenant and application namespaces. The secret is then marked immutable, which leads to problems if the certificate is updated later or if tenant SANs are added asynchronously to the cert (see #923).

This PR switches to a model of one cert object per tenant + one for the application namespace. The downside is that this means Ntenants+1 certs will be created with LetsEncrypt when the cluster is first created, when previously it was one cert for everyone. However it avoids the need to copy certificates across namespaces and keeps each tenant on a granular certificate. It also removes the prior immutable setting, allowing certs to be renewed automatically when they're approaching their expiration.

An alternate strategy would be to create a single `*.cluster.domain.io` wildcard cert to be shared by everybody, thereby reducing the number of interactions with LetsEncrypt. However we would still likely want to remove the immutable flag on the secret in order to allow expiration updates.

Signed-off-by: Nick Parker <nick@opstrace.com>

* controller: need to specify type=ClusterIssuer in certificate objects

Signed-off-by: Nick Parker <nick@opstrace.com>

* controller: skip calculating diff if debug is not enabled

Signed-off-by: Nick Parker <nick@opstrace.com>

* controller: avoid setting annotation to undefined value

Signed-off-by: Nick Parker <nick@opstrace.com>

* controller: subscribe to v1 format CRD stream instead of v1beta1

Signed-off-by: Nick Parker <nick@opstrace.com>