cli: aws infra setup: set ec2 instance metadata
This is executed as part of an upgrade, in particular
as part of the 'infrastructure upgrade'.

When this runs, EC2 instance metadata is mutated
so that HttpPutResponseHopLimit is set from 1 (the
default to 2). This is a pragmatic first step
#1383.

Note that this per-instance mutation is not
persistent: goes away when an intance goes away
in an auto-scaling group.

To persist this kind of configuration change
the underlying launch configuration needs to
be mutated (more context in #1383).

Signed-off-by: Dr. Jan-Philip Gehrcke <jp@opstrace.com>