cortex: remap 'retryable' 5xx errors to non-retryable errors
When cortex is backing up (e.g. full distributor queues) it can return 500 errors.
If clients retry-loop on these errors, it can lead to nobody being able to catch up.

Note that this is separate from the existing 429->503 mapping, which stays as-is.

See also https://github.com/opstrace/opstrace/issues/1409

Signed-off-by: Nick Parker <nick@opstrace.com>